requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.7
oauthlib==2.1.0
py-trello==0.14.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.24.1
